requirements.txt
Django
psycopg_binary
djangorestframework